taprio: Add support for cycle_time and cycle_time_extension
This allows a cycle-time and a cycle-time-extension to be specified.

Specifying a cycle-time will truncate that cycle, so when that instant
is reached, the cycle will start from its beginning.

A cycle-time-extension may cause the last entry of a cycle, just
before the start of a new schedule (its base-time) to be extended by
at maximum "cycle-time-extension" nanoseconds. So there's no too short
cycles.

Example:

tc qdisc change dev veth0 parent root handle 100 taprio \
	      sched-entry S 0x1 1000000 \
	      sched-entry S 0x0 2000000 \
	      sched-entry S 0x1 3000000 \
	      sched-entry S 0x0 4000000 \
	      cycle-time-extension 10000000 \
	      base-time 8000000000000 \

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@intel.com>